+++ example/lib/splash/splash_screen_model.dart
@@ -1,0 +1,6 @@
+import 'package:flutter_arhitecture_helper/presentation/ui/mvvm/stateful/base_model.dart';
+import 'package:flutter_arhitecture_helper/presentation/ui/mvvm/utils/base_model_utils.dart';
+
+class SplashScreenModel extends BaseModel {
+  OnCallCommand click = new OnCallCommand();
+}