+++ CHANGELOG.md
@@ -1,3 +1,43 @@
-## 0.0.1
+# File with heading
 
-* TODO: Describe initial release.
+## [0.4.0, 0.4.1] - 07.05.2019
+
+* Implement BaseApp as MVVM
+* Added View Events (paused, suspending, resumed etc)
+
+## [0.3.8, 0.3.9] - 04.04.2019
+
+* Added Base App parameters
+
+## [0.3.7] - 01.04.2019
+
+* Implement Base App
+* Implement loading widget
+
+## [0.3.6] - 27.02.2019
+
+* Fix navigation.
+
+## [0.3.5] - 18.02.2019
+
+* Refactoring.
+
+## [0.3.4] - 11.02.2019
+
+* Fix package to use with stateful and stateless widgets.
+
+## [0.3.3] - 02.02.2019
+
+* Fix view created action.
+
+## [0.3.0] - 25.01.2019
+
+* Release.
+
+## [0.0.2] - 24.01.2019
+
+* Fixes.
+
+## [0.0.1] - 22.01.2019
+
+* First Alpha version of package.